refactor: Moved editmodal into UserRow and UserFilters into UserTable.
